--- src/main/java/ru/yandex/filmorate/tests/FilmoRateApplicationTests.java
@@ -1,30 +1,0 @@
-package ru.yandex.filmorate.tests;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.context.SpringBootTest;
-import ru.yandex.filmorate.model.User;
-import ru.yandex.filmorate.storage.user.UserDbStorage;
-
-
-import java.util.Optional;
-
-@SpringBootTest
-@AutoConfigureTestDatabase
-@RequiredArgsConstructor(onConstructor_ = @Autowired)
-class FilmoRateApplicationTests {
-    private final UserDbStorage userStorage;
-
-
-    public void testFindUserById() {
-
-     /*  Optional<User> userOptional = userStorage.get(1);
-
-        assertThat(userOptional)
-                .isPresent()
-                .hasValueSatisfying(user ->
-                        assertThat(user).hasFieldOrPropertyWithValue("id", 1)
-                );*/
-    }
-}
