+++ src/test/java/ru/yandex/filmorate/FilmorateApplicationTests.java
@@ -1,79 +1,32 @@
 package ru.yandex.filmorate;
 
+import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.context.SpringBootTest;
-import ru.yandex.filmorate.controllers.FilmController;
-import ru.yandex.filmorate.controllers.UserController;
-import ru.yandex.filmorate.model.Film;
 import ru.yandex.filmorate.model.User;
+import ru.yandex.filmorate.storage.user.UserStorage;
 
-import java.time.LocalDate;
+import java.util.Optional;
 
 @SpringBootTest
+@AutoConfigureTestDatabase
+@RequiredArgsConstructor(onConstructor_ = @Autowired)
 class FilmorateApplicationTests {
-
-    /*@Test
-    void checkFilmWithNoName() {
-        Film film = Film.builder()
-                .name("")
-                .description("test")
-                .releaseDate(LocalDate.of(2000, 12, 12))
-                .duration(199)
-                .build();
-        FilmController filmController = new FilmController();
-        try {
-            filmController.createFilm(film);
-        } catch (ValidationException | SimilarException e) {
-            assertEquals("Name is needed", e.getMessage());
-        }
-    }
+    //private final UserStorage userDbStorage;
 
     @Test
-    void checkFilmDate() {
-        Film film = Film.builder()
-                .name("test")
-                .description("test")
-                .releaseDate(LocalDate.of(1000, 12, 12))
-                .duration(199)
-                .build();
-        FilmController filmController = new FilmController();
-        try {
-            filmController.createFilm(film);
-        } catch (ValidationException | SimilarException e) {
-            assertEquals("Date of release must after 27.12.1985", e.getMessage());
-        }
+    public void testFindUserById() {
+
+        //Optional<User> userOptional = userDbStorage.get(1);
+
+
+        System.out.println();
+        /*assertThat(userOptional)
+                .isPresent()
+                .hasValueSatisfying(user ->
+                        assertThat(user).hasFieldOrPropertyWithValue("id", 1)
+                );*/
     }
-
-    @Test
-    void checkUserEmail() {
-        User user = User.builder()
-                .email("@")
-                .login("marat")
-                .name("marat")
-                .birthday(LocalDate.of(3000, 12, 12))
-                .build();
-        UserController userController = new UserController();
-        try {
-            userController.createUser(user);
-        } catch (ValidationException | SimilarException e) {
-            assertEquals("Bad email. Example xxx@xxx.ru", e.getMessage());
-        }
-    }
-
-    @Test
-    void checkUserBirthday() {
-        User user = User.builder()
-                .email("marat@yandex.ru")
-                .login("marat")
-                .name("marat")
-                .birthday(LocalDate.of(3000, 12, 12))
-                .build();
-        UserController userController = new UserController();
-        try {
-            userController.createUser(user);
-        } catch (ValidationException | SimilarException e) {
-            assertEquals("You are from future", e.getMessage());
-        }
-    }*/
-
 }
